+++ protected/humhub/modules/file/messages/pt_br/widgets_UploadInput.php
@@ -1,5 +1,5 @@
 <?php
-return array (
+return [
   'Sorry, you can only upload up to {n,plural,=1{# file} other{# files}} at once.' => 'Desculpe, você só pode enviar {n,plural,=1{# arquivo} other{# arquivos}} de uma vez.',
-  'This upload field only allows a maximum of {n,plural,=1{# file} other{# files}}.' => 'Esse campo só permite o envio de até {n,plural,=1{# arquivo} other{# arquivos}}.',
-);
+  'This upload field only allows a maximum of {n,plural,=1{# file} other{# files}}.' => 'Esse campo só permite o envio de até {n,plural,=1{# file} other{# files}}.',
+];